Moved code into new classification class
