+++ src/DotNetCensus.Core/Census.cs
@@ -10,7 +10,7 @@ namespace DotNetCensus.Core
             List<FrameworkSummary> frameworkSummary = new();
             foreach (Project project in projects)
             {
-                project.Family = ProjectScanning.GetFrameworkFamily(project.FrameworkCode);
+                project.Family = Classification.GetFrameworkFamily(project.FrameworkCode);
                 if (string.IsNullOrEmpty(project.FrameworkCode) == true)
                 {
                     project.FrameworkCode = "(Unknown framework)";
@@ -26,7 +26,7 @@ namespace DotNetCensus.Core
                     {
                         Framework = project.FrameworkName,
                         FrameworkFamily = project.Family,
-                        Status = ProjectScanning.GetStatus(project.FrameworkCode),
+                        Status = Classification.GetStatus(project.FrameworkCode),
                         Count = 1 //it's the first time, start with a count of 1
                     });
                 }
