+++ src/DotNetCensus.Core/Models/Project.cs
@@ -9,5 +9,6 @@
         public string? Family { get; set; }
         public string? Language { get; set; }
         public string? Status { get; set; }
+        public string? Content { get; set; }
     }
 }
